moving DAU / DHU / N3U from types to dnssec
