+++ dnsext-dnssec/DNS/SEC/Types.hs
@@ -4,8 +4,7 @@
 {-# LANGUAGE TransformListComp #-}
 
 module DNS.SEC.Types (
-    addResourceDataForDNSSEC
-  , TYPE (
+    TYPE (
     RRSIG
   , DS
   , NSEC
@@ -375,16 +374,3 @@ getNsecTypes len = concat <$> sGetMany "NSEC type bitmap" len getbits
         blkTypes (bitOffset, byte) =
             [ toTYPE $ fromIntegral $ bitOffset + i |
               i <- [0..7], byte .&. bit (7-i) /= 0 ]
-
-----------------------------------------------------------------
-
-addResourceDataForDNSSEC :: InitIO ()
-addResourceDataForDNSSEC = do
-  extendRR DS      "DS"      (Proxy :: Proxy RD_DS)
-  extendRR DNSKEY  "DNSKEY"  (Proxy :: Proxy RD_DNSKEY)
-  extendRR CDS     "CDS"     (Proxy :: Proxy RD_CDS)
-  extendRR CDNSKEY "CDNSKEY" (Proxy :: Proxy RD_CDNSKEY)
-  extendRR RRSIG   "RRSIG"   (Proxy :: Proxy RD_RRSIG)
-  extendRR NSEC    "NSEC"    (Proxy :: Proxy RD_NSEC)
-  extendRR NSEC3   "NSEC3"   (Proxy :: Proxy RD_NSEC3)
-  extendRR NSEC3PARAM "NSEC3PARAM" (Proxy :: Proxy RD_NSEC3PARAM)
