+++ dnsext-types/DNS/Types/EDNS.hs
@@ -9,9 +9,6 @@ module DNS.Types.EDNS (
   , OptCode (
     OptCode
   , NSID
-  , DAU
-  , DHU
-  , N3U
   , ClientSubnet
   )
   , fromOptCode
@@ -23,14 +20,8 @@ module DNS.Types.EDNS (
   , encodeOData
   , OData(..)
   , OD_NSID(..)
-  , OD_DAU(..)
-  , OD_DHU(..)
-  , OD_N3U(..)
   , OD_ClientSubnet(..)
   , od_nsid
-  , od_dau
-  , od_dhu
-  , od_n3u
   , od_clientSubnet
   , od_ecsGeneric
   , od_unknown
@@ -134,14 +125,6 @@ toOptCode = OptCode
 pattern NSID :: OptCode
 pattern NSID  = OptCode 3
 
--- | DNSSEC algorithm support (RFC6975, section 3)
-pattern DAU  :: OptCode
-pattern DAU   = OptCode 5
-pattern DHU  :: OptCode
-pattern DHU   = OptCode 6
-pattern N3U  :: OptCode
-pattern N3U   = OptCode 7
-
 -- | Client subnet (RFC7871)
 pattern ClientSubnet :: OptCode
 pattern ClientSubnet = OptCode 8
@@ -166,9 +149,6 @@ insertOptShowDict (OptCode w) name dict = IM.insert i name dict
 defaultOptShowDict :: OptShowDict
 defaultOptShowDict =
     insertOptShowDict NSID "NSID"
-  $ insertOptShowDict DAU  "DAU"
-  $ insertOptShowDict DHU  "DHU"
-  $ insertOptShowDict N3U  "N3U"
   $ insertOptShowDict ClientSubnet "ClientSubnet"
    IM.empty
 
@@ -198,9 +178,6 @@ insertOptReadDict o name dict = M.insert name o dict
 defaultOptReadDict :: OptReadDict
 defaultOptReadDict =
     insertOptReadDict NSID "NSID"
-  $ insertOptReadDict DAU  "DAU"
-  $ insertOptReadDict DHU  "DHU"
-  $ insertOptReadDict N3U  "N3U"
   $ insertOptReadDict ClientSubnet "ClientSubnet"
    M.empty
 
@@ -266,57 +243,6 @@ instance OptData OD_NSID where
 
 od_nsid :: Opaque -> OData
 od_nsid = toOData . OD_NSID
-
----------------------------------------------------------------
-
--- | DNSSEC Algorithm Understood (RFC6975).  Client to server.
--- (array of 8-bit numbers). Lists supported DNSKEY algorithms.
-newtype OD_DAU = OD_DAU [Word8] deriving (Eq)
-
-instance Show OD_DAU where
-    show (OD_DAU as) = _showAlgList "DAU" as
-
-instance OptData OD_DAU where
-    optDataCode _ = DAU
-    encodeOptData (OD_DAU as) = putODWords (fromOptCode DAU) as
-    decodeOptData _ len = OD_DAU <$> getNOctets len
-
-od_dau :: [Word8] -> OData
-od_dau a = toOData $ OD_DAU a
-
----------------------------------------------------------------
-
--- | DS Hash Understood (RFC6975).  Client to server.
--- (array of 8-bit numbers). Lists supported DS hash algorithms.
-newtype OD_DHU = OD_DHU [Word8] deriving (Eq)
-
-instance Show OD_DHU where
-    show (OD_DHU hs)    = _showAlgList "DHU" hs
-
-instance OptData OD_DHU where
-    optDataCode _ = DHU
-    encodeOptData (OD_DHU hs) = putODWords (fromOptCode DHU) hs
-    decodeOptData _ len = OD_DHU <$> getNOctets len
-
-od_dhu :: [Word8] -> OData
-od_dhu a = toOData $ OD_DHU a
-
----------------------------------------------------------------
-
--- | NSEC3 Hash Understood (RFC6975).  Client to server.
--- (array of 8-bit numbers). Lists supported NSEC3 hash algorithms.
-newtype OD_N3U = OD_N3U [Word8] deriving (Eq)
-
-instance Show OD_N3U where
-    show (OD_N3U hs)    = _showAlgList "N3U" hs
-
-instance OptData OD_N3U where
-    optDataCode _ = N3U
-    encodeOptData (OD_N3U hs) = putODWords (fromOptCode N3U) hs
-    decodeOptData _ len = OD_N3U <$> getNOctets len
-
-od_n3u :: [Word8] -> OData
-od_n3u a = toOData $ OD_N3U a
 
 ---------------------------------------------------------------
 
@@ -454,9 +380,6 @@ od_unknown code o = toOData $ OD_Unknown code o
 
 ---------------------------------------------------------------
 
-_showAlgList :: String -> [Word8] -> String
-_showAlgList nm ws = nm ++ " " ++ intercalate "," (map show ws)
-
 _showNSID :: OD_NSID -> String
 _showNSID (OD_NSID nsid) = "NSID "
                         ++ b16encode bs
@@ -472,14 +395,6 @@ _showECS family srcBits scpBits address =
                 ++ " " ++ show scpBits ++ " " ++ address
 
 ---------------------------------------------------------------
-
--- | Encode EDNS OPTION consisting of a list of octets.
-putODWords :: Word16 -> [Word8] -> SPut
-putODWords code ws =
-     mconcat [ put16 code
-             , putInt16 $ length ws
-             , mconcat $ map put8 ws
-             ]
 
 -- | Encode an EDNS OPTION byte string.
 putODBytes :: Word16 -> Opaque -> SPut
